LayoutTests/platform/wpe/media/accessiblity-describes-video-expected.txt


EVENT(canplaythrough)
EXPECTED (video.audioTracks.length == '2') OK
EXPECTED (video.audioTracks[0].enabled == 'false'), OBSERVED 'true' FAIL
EXPECTED (video.audioTracks[1].enabled == 'true'), OBSERVED 'false' FAIL
END OF TEST
